Add files via upload
First upload. 

Currently, game is bugged. At certain iteration, the game becomes unstable and the neural network controlled point is terminated at round start for every subsequent iteration. There is also no logic for the input of the neural network point. The plan is for the point to 'see' the location and velocity of the closest n collision points.